Refactor 'What Links Here'

+++ app/controllers/references_controller.rb
@@ -65,8 +65,13 @@ class ReferencesController < ApplicationController
   def destroy
     reference = find_reference
 
-    # Grab key before reference author names are deleted.
-    activity_parameters = { name: reference.keey }
+    if References::WhatLinksHere.new(reference).any?
+      redirect_to reference_what_links_here_path(reference),
+        alert: "Other records refer to this reference, so it can't be deleted."
+      return
+    end
+
+    activity_parameters = { name: reference.keey } # Grab key before reference author names are deleted.
 
     if reference.destroy
       reference.create_activity :destroy, current_user, parameters: activity_parameters
@@ -83,26 +88,7 @@ class ReferencesController < ApplicationController
     end
 
     def reference_params
-      params.require(:reference).permit(
-        :author_names_string,
-        :author_names_suffix,
-        :bolton_key,
-        :citation_year,
-        :date,
-        :doi,
-        :editor_notes,
-        :journal_name,
-        :nesting_reference_id,
-        :online_early,
-        :pagination,
-        :public_notes,
-        :publisher_string,
-        :series_volume_issue,
-        :stated_year,
-        :taxonomic_notes,
-        :title,
-        document_attributes: [:id, :file, :url]
-      )
+      params.require(:reference).permit(ReferenceForm::PERMITTED_PARAMS)
     end
 
     # NOTE: This is so references can be converted between types.
